Make ScorePerformanceProcessor use the new PerformanceProcessor
