+++ osu.Server.Queues.ScoreStatisticsProcessor/Processors/ScorePerformanceProcessor.cs
@@ -1,16 +1,8 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Linq;
 using Dapper;
 using MySqlConnector;
-using osu.Game.Beatmaps.Legacy;
-using osu.Game.Rulesets;
-using osu.Game.Rulesets.Difficulty;
-using osu.Game.Rulesets.Mania.Mods;
-using osu.Game.Rulesets.Mods;
-using osu.Game.Rulesets.Osu.Mods;
-using osu.Game.Scoring;
 using osu.Server.Queues.ScoreStatisticsProcessor.Models;
 
 namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
@@ -23,88 +15,24 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
 
         public void ApplyToUserStats(SoloScoreInfo score, UserStats userStats, MySqlConnection conn, MySqlTransaction transaction)
         {
-            Ruleset ruleset = ScoreStatisticsProcessor.AVAILABLE_RULESETS.Single(r => r.RulesetInfo.OnlineID == score.ruleset_id);
-            Mod[] mods = score.mods.Select(m => m.ToMod(ruleset)).ToArray();
-            ScoreInfo scoreInfo = score.ToScoreInfo(mods);
+            // .Wait() in this class should be safe, since the queue processor runs under its own (non-TPL) task scheduler.
+            var dbInfo = LegacyDatabaseHelper.GetRulesetSpecifics(score.ruleset_id);
+            var processor = PerformanceProcessor.CreateAsync(conn, transaction).Result;
 
-            double performance = computePerformance(ruleset, mods, scoreInfo, conn, transaction);
+            processor.ProcessScoreAsync(score, conn, transaction).Wait();
+            processor.UpdateUserStatsAsync(userStats, score.ruleset_id, conn, transaction).Wait();
 
-            conn.Execute($"INSERT INTO {SoloScorePerformance.TABLE_NAME} (`score_id`, `pp`) VALUES (@ScoreId, @PP) ON DUPLICATE KEY UPDATE `pp` = @PP", new
+            int warnings = conn.QuerySingleOrDefault<int>($"SELECT `user_warnings` FROM {dbInfo.UsersTable} WHERE `user_id` = @UserId", new
             {
-                ScoreId = score.id,
-                PP = performance
+                UserId = userStats.user_id
             }, transaction);
+
+            if (warnings > 0)
+                userStats.rank_score = 0;
         }
 
         public void ApplyGlobal(SoloScoreInfo score, MySqlConnection conn)
         {
-        }
-
-        private double computePerformance(Ruleset ruleset, Mod[] mods, ScoreInfo score, MySqlConnection conn, MySqlTransaction transaction)
-        {
-            if (!AllModsValidForPerformance(mods))
-                return 0;
-
-            var beatmap = conn.QuerySingle<Beatmap>("SELECT * FROM osu_beatmaps WHERE beatmap_id = @BeatmapId", new
-            {
-                BeatmapId = score.BeatmapInfo.OnlineID
-            }, transaction);
-
-            // Todo: We shouldn't be using legacy mods, but this requires difficulty calculation to be performed in-line.
-            LegacyMods legacyModValue = LegacyModsHelper.MaskRelevantMods(ruleset.ConvertToLegacyMods(mods), score.RulesetID != beatmap.playmode, score.RulesetID);
-
-            var rawDifficultyAttribs = conn.Query<BeatmapDifficultyAttribute>(
-                "SELECT * FROM osu_beatmap_difficulty_attribs WHERE beatmap_id = @BeatmapId AND mode = @RulesetId AND mods = @ModValue", new
-                {
-                    BeatmapId = score.BeatmapInfo.OnlineID,
-                    RulesetId = score.RulesetID,
-                    ModValue = (uint)legacyModValue
-                }, transaction).ToArray();
-
-            DifficultyAttributes attributes = LegacyRulesetHelper.CreateDifficultyAttributes(score.RulesetID);
-            attributes.FromDatabaseAttributes(rawDifficultyAttribs.ToDictionary(a => (int)a.attrib_id, a => (double)a.value), beatmap.ToAPIBeatmap());
-
-            return ruleset.CreatePerformanceCalculator()?.Calculate(score, attributes).Total ?? 0;
-        }
-
-        /// <summary>
-        /// Checks whether all mods in a given array are valid to give PP for.
-        /// </summary>
-        public static bool AllModsValidForPerformance(Mod[] mods)
-        {
-            foreach (var m in mods)
-            {
-                switch (m)
-                {
-                    case ManiaModHardRock:
-                    case ManiaModKey1:
-                    case ManiaModKey2:
-                    case ManiaModKey3:
-                    case ManiaModKey10:
-                        return false;
-
-                    case ModEasy:
-                    case ModNoFail:
-                    case ModHalfTime:
-                    case ModSuddenDeath:
-                    case ModPerfect:
-                    case ModHardRock:
-                    case ModDoubleTime:
-                    case ModHidden:
-                    case ModFlashlight:
-                    case ModMuted:
-                    case ModClassic:
-                    case OsuModSpunOut:
-                    case ManiaKeyMod:
-                    case ManiaModMirror:
-                        continue;
-
-                    default:
-                        return false;
-                }
-            }
-
-            return true;
         }
     }
 }
